fixes for Windows
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@37110 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 9b164d2afb93103f4e456f5b297c2891357f58c0

+++ LHCb/Event/EventPacker/src/PackMCParticle.cpp
@@ -1,4 +1,4 @@
-// $Id: PackMCParticle.cpp,v 1.1.1.1 2008-11-18 17:12:59 ocallot Exp $
+// $Id: PackMCParticle.cpp,v 1.2 2008-11-19 13:00:55 cattanem Exp $
 // Include files 
 
 // from Gaudi
@@ -58,7 +58,7 @@ StatusCode PackMCParticle::execute() {
     newPart.px   = pack.energy( part->momentum().px() );
     newPart.py   = pack.energy( part->momentum().py() );
     newPart.pz   = pack.energy( part->momentum().pz() );
-    newPart.mass = part->virtualMass();
+    newPart.mass = (float)part->virtualMass();
     newPart.PID  = part->particleID().pid();
     newPart.originVertex = pack.reference( out, 
                                            part->originVertex()->parent(),
